Introduction of new daemon atopacctd. This daemon switches on process accounting and transfers every accounting record to an accounting shadow file. The source accounting file will regularly be truncated, while the shadow files are written with a limited size in a queued way. Non-used shadow files are deleted regularly.

+++ acctproc.h
@@ -21,11 +21,11 @@
 ** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 ** See the GNU General Public License for more details.
 */
-int 	acctswon(void);
-void	acctswoff(void);
-int 	acctprocnt(void);
-int 	acctphotoproc(struct tstat *, int);
-void 	acctrepos(unsigned int);
+int 		acctswon(void);
+void		acctswoff(void);
+unsigned long 	acctprocnt(void);
+int 		acctphotoproc(struct tstat *, int);
+void 		acctrepos(unsigned int);
 
 /*
 ** maximum number of records to be read from process accounting file
